% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sino_null.R
\name{sino_null}
\alias{sino_null}
\title{Simulated Normality (SiNo) Test}
\usage{
sino_null(size, prec = 1024, dlim = 10, sims = 1000)
}
\arguments{
\item{size}{A vector containing the data}

\item{prec}{The number of equally spaced points at which the density
is to be estimated using \code{stats::density}}

\item{dlim}{The value of \code{cut} in \code{stats::density}, which
dictates how the density values at extremes are estimated (values dropped
to approximately 0 at the extremes)}

\item{sims}{The number of Monte Carlo simulations used to
generate the null distribution}
}
\value{
An object of class \code{density}
}
\description{
This generates the null distirbution for the SiNo Test.
}
\examples{
sino_null(550)

}
\seealso{
\code{\link[=sino_test]{sino_test()}}
}
