% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sino_test.R
\name{sino_test}
\alias{sino_test}
\title{Simulated Normality (SiNo) Test}
\usage{
sino_test(
  x,
  simulations = 1000,
  precision = 1024,
  def_lim = 10,
  ref_dist = NULL,
  sig_level = 0.05,
  show_message = TRUE
)
}
\arguments{
\item{x}{A vector containing the data}

\item{simulations}{The number of Monte Carlo simulations used to
generate the null distribution when \code{ref_dist = NULL}}

\item{precision}{The number of equally spaced points at which the density
is to be estimated using \code{stats::density}}

\item{def_lim}{The value of \code{cut} in \code{stats::density}, which
dictates how the density values at extremes are estimated (values dropped
to approximately 0 at the extremes)}

\item{ref_dist}{The null distribution, either as an object of class
\code{density} (i.e., kernel density estimate) or a cumulative distribution
function (i.e., an object of class \code{function})}

\item{sig_level}{The level of significance}

\item{show_message}{If \code{TRUE}, results will be printed out}
}
\value{
A named list containing the following components: \cr
\itemize{
\item{\code{data_name}}{ - The name of the dataset provided}
\item{\code{data}}{ - The dataset placed under the test}
\item{\code{statistic}}{ - The test statistic}
\item{\code{p_value}}{ - The p-value}
\item{\code{simulations}}{ - The value passed to \code{simulations} when
the function was called}
\item{\code{precision}}{ - The value passed to \code{precision} when
the function was called}
\item{\code{def_lim}}{ - The value passed to \code{def_lim} when
the function was called}
\item{\code{ref_dist}}{ - The null distribution used, either as a
density object or a cumulative distribution function}
}
}
\description{
The Simulated Normality (or SiNo) Test is a Lilliefors-like
Kernel Density Estimate Test for Normality. It uses the integral of the
L1-distance over the reals as the test statistic. The null distribution
is estimated using Monte Carlo simulations. The realized sample
is first standardized (z-score) and then, the kernel density estimate
is obtained based on the standardized sample. This kernel density
estimate is then compared with the standard normal density.
}
\note{
If \code{ref_dist = NULL}, the results of the test will be
non-deterministic for sample sizes less than 5 or greater than 2000. This is
because the null distribution is generated every time using Monte Carlo
simulations for such cases. Pre-computed null distributions are available
only for samples with size between 5 and 2000 (inclusive). Thus, running the
test on the same dataset more than once with \code{ref_dist = NULL} may
yield different test statistics (but they are expected to be close). You
may generate your own estimates for the null distribution using
\code{SinoTest::sino_null}.
}
\examples{
sino_test(rnorm(1000))
sino_test(rexp(100), simulations = 500)

}
\seealso{
\code{\link[=sino_null]{sino_null()}}
}
\keyword{normality}
